add retry mechanism; simplify adding workers
mod_push_gcm now has a retry mechanism. It examines the response from
GCM and decides whether the request was successful or - if not - if the
error is recoverable. If recoverable the data needed for
crafting a push notification is stored in a list (only one entry per user).
The retry list is processed every ?RETRY_PERIOD milliseconds.
In order to store only one entry per user the dispatch call needed an
extra argument: BareUser. All backends have been adapted.

The process of starting workers has been simplified. The worker
processes are now named after the backend's register_host and type.
We no longer avoid workers with duplicate auth_data. There's no
significant advantage of doing so and the use case probably is very seldom.