push on all incoming stanzas and store message stanzas
mod_push does not filter by stanza type any more. All incoming stanzas
now cause a push notification to be sent (with or without an updated
payload). If a client does not want to receive certain stanzas it has to
make use of other XEPs such as Client State Indication.
In order to not lose messages in case of a server restart mod_push
stores message stanzas on disk. Iq stanzas are not stored because it
does not make sense to reply to iqs in a new session. Presence stanzas
are not stored either as they are generated newly when starting a new
session.